+++ fcxref/find/make_find.py
@@ -1,3 +1,4 @@
+import logging
 import re
 from typing import Callable, Dict, List
 from xml.etree.ElementTree import Element
@@ -8,12 +9,15 @@ from .reference import Reference
 
 __all__ = ['make_find']
 
+logger = logging.getLogger(__name__)
 
 def make_find(find_root_by_document_path: Callable[[str], Dict[str, Element]]):
     def find(base_path: str, query: Query) -> List[Reference]:
         references = []
         root_by_document_path = find_root_by_document_path(base_path)
+        logger.debug(f'Finding references in base path {base_path}')
         for document_path, root in root_by_document_path.items():
+            logger.debug(f'Checking document {document_path}')
             query_pattern = re.compile(query.to_regex())
             references_in_document = find_references_in_root(
                 document_path, root, query_pattern)